doctor command
