+++ src/commands/mod.rs
@@ -2,6 +2,7 @@ pub mod activate;
 pub mod alias;
 pub mod completions;
 pub mod deactivate;
+pub mod doctor;
 pub mod info;
 pub mod install;
 pub mod list;
