+++ src/main.rs
@@ -9,7 +9,6 @@ pub mod version;
 use clap::{CommandFactory, Parser};
 use clap_complete::Shell;
 use clap_derive::{Parser, Subcommand};
-use colored::Colorize;
 use meta::DvmMeta;
 use utils::dvm_root;
 
@@ -100,6 +99,9 @@ enum Commands {
   Activate,
   #[clap(about = "Deactivate Dvm")]
   Deactivate,
+
+  #[clap(about = "Fixing dvm specific environment variables and other issues")]
+  Doctor,
 }
 
 #[derive(Subcommand)]
@@ -126,18 +128,6 @@ pub fn main() {
   let cli = Cli::parse();
   let mut meta = DvmMeta::new();
 
-  // TODO(CGQAQ): Move this into the `doctor` command
-  // Init enviroments if need
-  // actually set DVM_DIR env var if not exist.
-  let home_path = dvm_root();
-  set_env::check_or_set("DVM_DIR", home_path.to_str().unwrap()).unwrap();
-  let path = set_env::get("PATH").unwrap();
-  let looking_for = deno_bin_path().parent().unwrap().to_str().unwrap().to_string();
-  if !path.contains(looking_for.as_str()) {
-    set_env::prepend("PATH", looking_for.as_str()).unwrap();
-    println!("{}", "Please restart your shell of choice to take effects.".red());
-  }
-
   let result = match cli.command {
     Commands::Completions { shell } => commands::completions::exec(&mut Cli::command(), shell),
     Commands::Info => commands::info::exec(),
@@ -149,6 +139,7 @@ pub fn main() {
     Commands::Alias { command } => commands::alias::exec(&mut meta, command),
     Commands::Activate => commands::activate::exec(&mut meta),
     Commands::Deactivate => commands::deactivate::exec(),
+    Commands::Doctor => commands::doctor::exec(&mut meta),
   };
 
   if let Err(err) = result {
